server/requirements.txt
flask
requests
PyPDF2
google-generativeai
python-dotenv